+++ src/Syncer.js
@@ -3,6 +3,7 @@
 const ping = require('ping');
 const {NodeSSH} = require('node-ssh')
 const {Rsync} = require('rsync2');
+const fs = require('fs');
 
 const timeString = function () {
     const now = new Date();
@@ -35,7 +36,7 @@ class Syncer {
         this.configurator = configurator
         this.messenger = messenger
         this.timeInterval = setInterval( async () => {
-            if (this.configurator.isCorrect() && this.isConnected() && this.uploadFileFailed.length) {
+            if (this.uploadFileFailed.length && this.configurator.isCorrect() && this.isConnected()) {
                 let outputArray = JSON.parse(JSON.stringify(this.uploadFileFailed))
                 this.uploadFileFailed = [];
 
@@ -57,17 +58,17 @@ class Syncer {
                 this.sftp.connect(this.configurator.config.sftpOptions).then(() => {
                     this.messenger.infoSuccess('Config load success: ' + this.configurator.config.rootPath)
                 }, (error)=> {
-                    console.warn("SFTP:" +  JSON.stringify(error))
+                    console.warn("SyncSFTP:" +  JSON.stringify(error))
                 })
                 if (this.configurator.config.useRsync) {
-                    this.rsync = new Rsync(this.configurator.config.rsyncPath)
+                    this.rsync = new Rsync({executable: this.configurator.config.rsyncPath})
 
                     this.rsync.exclude(this.configurator.config.rsyncExclude.length ? this.configurator.config.rsyncExclude: this.configurator.config.ignorePatterns);
                     this.rsync.shell(`${this.configurator.config.sshPath} -p 22`).setFlags('zarv')
                 }
             }
         }).catch((e) => {
-            console.warn("SFTP:" +  JSON.stringify(e))
+            console.warn("SyncSFTP:" +  JSON.stringify(e))
             this.messenger.error('Can\'t connect to server')
         })
     }
@@ -132,7 +133,7 @@ class Syncer {
             this.messenger.infoSuccess(time + ' Succesfully uploaded ' + filename)
         }).catch((error) => {
 
-            console.warn("SFTP:" +  JSON.stringify(error))
+            console.warn("SyncSFTP:" +  JSON.stringify(error))
             if (error.code == 12) {
                 let parent = destination
                 parent = parent.replace(/[^/]+$/, '')
@@ -159,6 +160,101 @@ class Syncer {
         this.sftp.execCommand('rm ' + destination,{ cwd:'/var/www' });
         this.sftp.execCommand('rm ' + destination + '/*',{ cwd:'/var/www' });
         this.sftp.execCommand('rmdir ' + destination, { cwd:'/var/www' });
+    }
+    async detectChanges() {
+        let text = ''
+        let rsync = new Rsync({executable: this.configurator.config.rsyncPath})
+        rsync.exclude(this.configurator.config.rsyncExclude.length ? this.configurator.config.rsyncExclude: this.configurator.config.ignorePatterns);
+        // rsync.shell(`${this.configurator.config.sshPath} -p 22`)
+        rsync.output((data) => {text += data.toString()},(data) => {console.warn(data.toString());} )
+        let destinationFirstPart = this.configurator.config.sftpOptions.username + '@' + this.configurator.config.sftpOptions.host + ':';
+        let destinationLastPart = this.configurator.config.remotePath;
+        rsync._sources = [];
+        rsync._sources.push(this.configurator.config.rootPath + '/')
+        rsync._destination = destinationFirstPart + destinationLastPart
+        rsync.set('dry-run')
+        rsync.set('no-perms')
+        rsync.set('no-owner')
+        rsync.set('no-group')
+        rsync.set('no-times')
+        rsync.set('delete')
+        rsync.set('recursive')
+        rsync.set('links')
+        rsync.set('checksum')
+        rsync.set('itemize-changes')
+        try {
+            return rsync.execute().then(exitCode => {
+                let lines = text.split(/\n/)
+                lines = lines.filter(item => !item.match(/^<f.\..+/))
+                let toUpload = lines.filter(item => item.match(/^<.+/)).map(item => item.replace(/^.+ /, ''))
+                let toDelete = lines.filter(item => item.match(/\*deleting.+/)).map(item => item.replace(/^.+ /, ''))
+                return {toUpload, toDelete}
+            }).catch((error) => {
+                console.warn("SyncSFTP:" +  JSON.stringify(error))
+            })
+        } catch (error) {
+            console.warn("SyncSFTP:" +  JSON.stringify(error))
+        }
+    }
+    async notifyAboutChanges() {
+        if (!this.configurator.isCorrect()) {
+            this.messenger.error('Config not load')
+            return false
+        }
+        if (!this.configurator.config.useRsync) {
+            this.messenger.error('For RSYNC Users only!')
+            return false
+        }
+        if (!this.isConnected()) {
+            this.messenger.error('Can\'t connect to server')
+            return false
+        }
+        try {
+            this.detectChanges().then(({toUpload, toDelete}) => {
+                for (let item of toUpload) {
+                    this.messenger.info('Need to upload -> ' + item)
+                }
+                for (let item of toDelete) {
+                    this.messenger.info('Need to delete -> ' + item)
+                }
+                this.messenger.infoSuccess('Total different size: ' + (toUpload.length + toDelete.length))
+            })
+        } catch (error) {
+            console.warn("SyncSFTP:" +  JSON.stringify(error))
+        }
+    }
+    async makeEqual() {
+        if (!this.configurator.isCorrect()) {
+            this.messenger.error('Config not load')
+            return false
+        }
+        if (!this.configurator.config.useRsync) {
+            this.messenger.error('For RSYNC Users only!')
+            return false
+        }
+        if (!this.isConnected()) {
+            this.messenger.error('Can\'t connect to server')
+            return false
+        }
+        try {
+            this.detectChanges().then(({toUpload, toDelete}) => {
+                for (let item of toUpload) {
+                    let destination = this.configurator.config.remotePath + '/' + item;
+                    destination = destination.replace(/\\/g, '/');
+                    destination = destination.replace(/\/\/+/g, '/');
+                    let time = timeString();
+                    let isDirectory = false;
+                    this.messenger.info(time + ' Uploading to -> ' + destination)
+                    this.uploadFile(destination, this.configurator.config.rootPath + '/' + item, isDirectory)
+                }
+                for (let item of toDelete) {
+                    this.deleteFile(this.configurator.config.remotePath + '/' + item)
+                    this.messenger.infoSuccess('Deleted: ' + item)
+                }
+            })
+        } catch (error) {
+            console.warn("SyncSFTP:" +  JSON.stringify(error))
+        }
     }
 }